Restructuration de l'arborescence du projet actuel pour qu'il ait le même que celui que j'avais fait et qui est la précédente version (ProjectV1)
