+++ web_app_frontend/src/View/ConnectionPage.vue
