+++ web_app_frontend/src/main.js
@@ -2,7 +2,7 @@ import './assets/main.css';
 
 import { createApp } from 'vue';
 import App from './App.vue';
-import router from './router';
+import router from './router/index.js';
 
 createApp(App)
   .use(router)
